Fetch data from api through language slice
